created teacher and course models, implemented CRUD operations for both models, created an abstract class person, and added a course registration feature
