+++ src/com/shreishta/training/exceptions/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package com.shreishta.training.services;
+
+public class UnauthorizedException extends Exception{
+    public UnauthorizedException(String message){
+        super(message);
+    }
+}
