+++ src/com/shreishta/training/models/Teacher.java
@@ -1,0 +1,4 @@
+package com.shreishta.training.models;
+
+public class Teacher {
+}
