+++ src/com/shreishta/training/services/CourseServiceImpl.java
@@ -1,0 +1,4 @@
+package com.shreishta.training.services;
+
+public class CourseServiceImpl {
+}
