+++ src/com/shreishta/training/services/TeacherServiceImpl.java
@@ -1,0 +1,4 @@
+package com.shreishta.training.services;
+
+public class TeacherServiceImpl {
+}
