+++ src/com/shreishta/training/services/courseService.java
@@ -1,0 +1,4 @@
+package com.shreishta.training.services;
+
+public interface courseService {
+}
