+++ src/com/shreishta/training/services/personInterface.java
@@ -1,0 +1,7 @@
+package com.shreishta.training.services;
+
+public interface person {
+    public int getRegNo();
+    public String getName();
+    public void setName(String name);
+}
