+++ src/com/shreishta/training/services/teacherService.java
@@ -1,0 +1,4 @@
+package com.shreishta.training.services;
+
+public interface teacherService {
+}
